enhancement: add image upload location mappings. support for de/fra canonical form. support for de/fra/2
